refactor: replace old components by new functional ones
