+++ src/components/Layout/HeaderOnlyLayout.jsx
@@ -1,4 +1,4 @@
-import HeaderTest from "../header/HeaderTest.jsx";
+import Header from "../header/Header.jsx";
 
 import s from "./Layout.module.scss";
 
@@ -6,7 +6,7 @@ function HeaderOnlyLayout({children}) {
 
   return (
     <>
-     <div className={s.header}><HeaderTest /></div>
+     <div className={s.header}><Header /></div>
      <div className={s.childrenWrap}>{children}</div>
     </>
   )
