+++ src/components/header/Header.module.scss
@@ -1,60 +1,22 @@
-/* All comments are in English only */
+// Media Queries variables
+$tablet: 37.5em;
+$laptop: 60em;
+$desktop: 80em;
 
 .headerContainer {
-  position: fixed;
-  flex-grow: 1;
-  top: 0;
-  width: 100%;
-  height: auto;
-  background-color: transparent;
-  /* Removed side paddings to not exceed 1152px container */
-  padding-left: 0;
-  padding-right: 0;
-  z-index: 3;
-  backdrop-filter: blur(3px);
-  background-color: rgba(255, 255, 255, 0.7);
-}
-
-.header {
   display: flex;
-  flex-direction: row;
   justify-content: space-between;
   align-items: center;
 
-  /* Constrain header to Figma width and center it */
+  padding: 0 1rem;
   width: 100%;
-  max-width: 1152px;
-  margin-left: auto;
-  margin-right: auto;
-  position: relative;
+  height: 6.25rem;
 
-  @media (min-width: 320px) and (max-width: 599.99px) {
-    padding: 0px 25px;
-    width: 100%;
-    height: 75px;
-    /* max-width keeps working but is not limiting on small screens */
-  }
+  background-color: transparent;
+  z-index: 3;
 
-  @media (min-width: 600px) and (max-width: 959.99px) {
-    margin-left: auto;
-    margin-right: auto;
-    padding: 10px 20px;
-    height: 81px;
-  }
-
-  @media (min-width: 960px) and (max-width: 1279.99px) {
-    margin-left: auto;
-    margin-right: auto;
-    padding: 10px 20px;
-    height: 81px;
-  }
-
-  @media (min-width: 1280px) and (max-width: 2200px) {
-    margin-left: auto;
-    margin-right: auto;
-    /* No extra side padding here to keep exact 1152px width */
-    height: 100px;
-    max-width: 1152px; /* explicit for desktop range */
+  @media (min-width: $desktop) {
+    padding: 0;
   }
 }
 
@@ -66,130 +28,70 @@
   user-select: none;
 }
 
-.menuIcon {
-  background: none;
-  border: none;
-  cursor: pointer;
-  padding: 0;
-  margin: 0;
-  outline: none;
-  box-shadow: none;
-  appearance: none;
-  -webkit-appearance: none;
-}
-
-.menuIcon:focus,
-.menuIcon:active {
-  outline: none;
-  box-shadow: none;
-  border: none;
-}
-
-.bookingButtonsWrapper {
-  margin-right: 25px;
+.btnsWrapper {
   display: flex;
-  flex-direction: row;
-  gap: 40px;
+  gap: 1rem;
 
-  // Smooth left offset when search is active for all 1280+ sizes
-  &.searchActive {
-    @media (min-width: 1280px) {
-      // Use CSS variable for dynamic offset with smooth transition
-      transform: translateX(var(--button-offset, 0px));
-      transition: transform 0.4s ease-in-out;
-    }
+  @media (min-width: 62.5em) {
+    gap: 2rem;
   }
-
-  // Animate the gap change smoothly
-  transition: gap 0.4s ease-in-out;
 }
 
-.rightPart {
+.iconsWrapper {
   display: flex;
-  flex-direction: row;
-  gap: 10px;
+  gap: 1rem;
   align-items: center;
-  position: relative;
+  justify-self: center;
+
+  @media (min-width: $laptop) {
+    gap: 2rem;
+  }
 }
 
-.searchWrapper {
-  position: relative;
-  display: flex;
-  align-items: center;
-}
-
-.searchInput {
-  position: absolute;
-  right: 38px;
-  width: 0;
-  opacity: 0;
-  padding: 8px 12px;
-  border: none;
-  border-radius: 5px;
-  font-size: 14px;
-  outline: none;
-  background-color: white;
-  color: #2f436e;
-  pointer-events: none;
-  transition: all 0.3s ease-in-out;
-  z-index: 10;
-  box-shadow: none;
-}
-
-.searchInput.open {
-  width: 200px;
-  opacity: 1;
-  pointer-events: auto;
-}
-
-.searchIconBtn {
-  background: none;
-  border: none;
-  cursor: pointer;
-  outline: none;
-  box-shadow: none;
-  display: flex;
-  align-items: center;
-  justify-content: center;
+.menuBtn {
   padding: 0;
   margin: 0;
-  appearance: none;
-  -webkit-appearance: none;
-}
-
-.searchIconBtn:focus,
-.searchIconBtn:active {
-  outline: none;
-  box-shadow: none;
   border: none;
+  background: none;
 }
 
-.searchIconBtn img {
-  width: 24px;
-  height: 24px;
-  outline: none;
-  border: none;
-  box-shadow: none;
-  appearance: none;
-  -webkit-appearance: none;
-  user-select: none;
-  pointer-events: none;
-}
+.menuImg {
+  width: 30px;
+  height: 25px;
 
-// Styles for BOOK DAY TOUR button
-.bookTourButton {
-  display: none;
+  @media (min-width: $laptop) {
+      width: 40px;
+      height: 30px
+    }
 }
 
 .btnCorrection {
-  font-size: 0.8rem;
+  display: none;
 
-  @media (min-width: 60rem) {
-  width: 11.25rem;
+  @media (min-width: 43.75em) { 
+    display: flex; 
+    font-size: 0.8rem;
+    width: 11.25rem;
   }
 
-  @media (min-width: 62.5rem) {
+  @media (min-width: 62.5em) {
   font-size: 1rem;
   width: 13.62rem;
+  height: 3.12rem;
   }
+
+  /* Only hide on desktop when isSearchOpern state is true */
+  @media (min-width: $desktop) {
+    &--hidden {
+      opacity: 0;
+      transform: translateX(-1rem) scale(.98);
+      max-width: 0;
+      padding-left: 0;
+      padding-right: 0;
+      margin-left: 0;
+      margin-right: 0;
+      visibility: hidden;
+      pointer-events: none;
+    }
+}
 }
