--- src/components/header/HeaderTest.module.scss
@@ -1,97 +1,0 @@
-// Media Queries variables
-$tablet: 37.5em;
-$laptop: 60em;
-$desktop: 80em;
-
-.headerContainer {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  padding: 0 1rem;
-  width: 100%;
-  height: 6.25rem;
-
-  background-color: transparent;
-  z-index: 3;
-
-  @media (min-width: $desktop) {
-    padding: 0;
-  }
-}
-
-.logoImage {
-  width: 179px;
-  height: 70px;
-  object-fit: contain;
-  cursor: pointer;
-  user-select: none;
-}
-
-.btnsWrapper {
-  display: flex;
-  gap: 1rem;
-
-  @media (min-width: 62.5em) {
-    gap: 2rem;
-  }
-}
-
-.iconsWrapper {
-  display: flex;
-  gap: 1rem;
-  align-items: center;
-  justify-self: center;
-
-  @media (min-width: $laptop) {
-    gap: 2rem;
-  }
-}
-
-.menuBtn {
-  padding: 0;
-  margin: 0;
-  border: none;
-  background: none;
-}
-
-.menuImg {
-  width: 30px;
-  height: 25px;
-
-  @media (min-width: $laptop) {
-      width: 40px;
-      height: 30px
-    }
-}
-
-.btnCorrection {
-  display: none;
-
-  @media (min-width: 43.75em) { 
-    display: flex; 
-    font-size: 0.8rem;
-    width: 11.25rem;
-  }
-
-  @media (min-width: 62.5em) {
-  font-size: 1rem;
-  width: 13.62rem;
-  height: 3.12rem;
-  }
-
-  /* Only hide on desktop when isSearchOpern state is true */
-  @media (min-width: $desktop) {
-    &--hidden {
-      opacity: 0;
-      transform: translateX(-1rem) scale(.98);
-      max-width: 0;
-      padding-left: 0;
-      padding-right: 0;
-      margin-left: 0;
-      margin-right: 0;
-      visibility: hidden;
-      pointer-events: none;
-    }
-}
-}
